+++ spring-api/src/main/java/eu/royalblackwater/api/account/AuthSessionRepository.java
@@ -4,9 +4,9 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthSessionRepository extends JpaRepository<AuthSessionEntity, Long> {
+public interface AuthSessionRepository extends JpaRepository<AuthSessionEntity, Integer> {
     Optional<AuthSessionEntity> findByTokenHash(String tokenHash);
     long deleteByTokenHash(String tokenHash);
-    long deleteByUserId(Long userId);
+    long deleteByUserId(Integer userId);
     long deleteByExpiresAtBefore(LocalDateTime threshold);
 }
