+++ spring-api/src/main/java/eu/royalblackwater/api/legal/LegalNoticeRepository.java
@@ -2,4 +2,4 @@ package eu.royalblackwater.api.legal;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LegalNoticeRepository extends JpaRepository<LegalNoticeEntity, Long> { }
+public interface LegalNoticeRepository extends JpaRepository<LegalNoticeEntity, Integer> { }
